add services layer, jwt token generation
