+++ openapi/auth-service.yml
@@ -1,7 +1,7 @@
-openapi: '3.0.2'
+openapi: "3.0.2"
 info:
   title: Auth API
-  version: '1.0'
+  version: "1.0"
 servers:
   - url: https://si-auth-sevice/api/v1/users
     description: production server
@@ -10,27 +10,21 @@ servers:
 
 components:
   schemas:
-      User:
-        type: object
-        required:
-          - "login"
-          - "password"
-        properties:
-          login:
-            type: string
-            format: email
-            example: client@mail.com
-          password:
-            type: string
-            format: password
-            example: passwd123
-          firstName:
-            type: string
-            example: Peter
-          lastName:
-            type: string
-            example: Quill
-  
+    User:
+      type: object
+      required:
+        - "login"
+        - "password"
+      properties:
+        login:
+          type: string
+          format: email
+          example: client@mail.com
+        password:
+          type: string
+          format: password
+          example: passwd123
+
   securitySchemes:
     bearerAuth:
       type: http
@@ -91,8 +85,8 @@ components:
 paths:
   /signup:
     post:
-      tags: 
-      - public
+      tags:
+        - public
       summary: Authenticate user credentials and return a JWT token
       operationId: signUp
       requestBody:
@@ -102,7 +96,7 @@ paths:
             schema:
               $ref: "#/components/schemas/User"
       responses:
-        '200':
+        "200":
           description: JWT token
           content:
             application/json:
@@ -112,14 +106,14 @@ paths:
                   token:
                     type: string
                     example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1ZWIxMmUxOTdlMDZhNzZjY2RlZmMxMjEiLCJpZCI6IjVlYjEyZTE5N2UwNmE3NmNjZGVmYzEyMSIsImlhdCI6MTU4ODc1ODE1N30.xR9H0STbFOpSkuGA9jHNZOJ6eS7umHHqKRhI807YT1Y"
-        '409':
-          $ref: '#/components/responses/AlreadyExists'
-        '500':
-          $ref: '#/components/responses/InternalError'
+        "409":
+          $ref: "#/components/responses/AlreadyExists"
+        "500":
+          $ref: "#/components/responses/InternalError"
   /signin:
     post:
-      tags: 
-      - public
+      tags:
+        - public
       summary: Authenticate user credentials and return a JWT token
       operationId: signIn
       requestBody:
@@ -133,12 +127,12 @@ paths:
                   type: string
                   format: email
                   example: client@mail.com
-                password: 
+                password:
                   type: string
                   format: password
                   example: passwd123
       responses:
-        '200':
+        "200":
           description: return JWT token
           content:
             application/json:
@@ -148,15 +142,15 @@ paths:
                   token:
                     type: string
                     example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1ZWIxMmUxOTdlMDZhNzZjY2RlZmMxMjEiLCJpZCI6IjVlYjEyZTE5N2UwNmE3NmNjZGVmYzEyMSIsImlhdCI6MTU4ODc1ODE1N30.xR9H0STbFOpSkuGA9jHNZOJ6eS7umHHqKRhI807YT1Y"
-        '400':
-          $ref: '#/components/responses/LoginPwdError'
-        '500':
-          $ref: '#/components/responses/InternalError'
-          
+        "400":
+          $ref: "#/components/responses/LoginPwdError"
+        "500":
+          $ref: "#/components/responses/InternalError"
+
   /{userId}/change-password:
     patch:
-      tags: 
-      - secure
+      tags:
+        - secure
       summary: change user password
       operationId: changePwd
       security:
@@ -174,31 +168,31 @@ paths:
             schema:
               type: object
               properties:
-                password: 
+                password:
                   type: string
                   format: password
                   example: passwd123
-                newPassword: 
+                newPassword:
                   type: string
                   format: password
                   example: f$MF(3EmG`#WRu&a
-                
+
       responses:
-        '204':
+        "204":
           description: User password changed successfully
-        '400':
-          $ref: '#/components/responses/LoginPwdError'
-        '401':
-          $ref: '#/components/responses/UnauthorizedError'
-        '404':
-          $ref: '#/components/responses/NotFoundError'
-        '500':
-          $ref: '#/components/responses/InternalError'
-  
+        "400":
+          $ref: "#/components/responses/LoginPwdError"
+        "401":
+          $ref: "#/components/responses/UnauthorizedError"
+        "404":
+          $ref: "#/components/responses/NotFoundError"
+        "500":
+          $ref: "#/components/responses/InternalError"
+
   /users/{userId}/change-role:
     patch:
-      tags: 
-      - secure
+      tags:
+        - secure
       summary: change user role
       operationId: changeRole
       security:
@@ -216,15 +210,15 @@ paths:
             schema:
               type: object
               properties:
-                newRole: 
+                newRole:
                   type: string
                   example: manager
       responses:
-        '204':
-          description:  User role changed successfully
-        '401':
-          $ref: '#/components/responses/UnauthorizedError'
-        '404':
-          $ref: '#/components/responses/NotFoundError'
-        '500':
-          $ref: '#/components/responses/InternalError'
+        "204":
+          description: User role changed successfully
+        "401":
+          $ref: "#/components/responses/UnauthorizedError"
+        "404":
+          $ref: "#/components/responses/NotFoundError"
+        "500":
+          $ref: "#/components/responses/InternalError"
