+++ sql/authdb.sql
@@ -1,5 +1,3 @@
-CREATE DATABASE authdb;
-
 CREATE TABLE roles (
   id SERIAL PRIMARY KEY,
   role VARCHAR(50) NOT NULL
